removed spring security and session and cookies adds

+++ src/main/java/com/matthewgeorgiev/project/MoviePreview/controller/UserController.java
@@ -1,6 +1,7 @@
 package com.matthewgeorgiev.project.MoviePreview.controller;
 
 import com.matthewgeorgiev.project.MoviePreview.model.User;
+import com.matthewgeorgiev.project.MoviePreview.service.AuthenticationService;
 import com.matthewgeorgiev.project.MoviePreview.service.UserService;
 import com.matthewgeorgiev.project.MoviePreview.utils.Helper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,6 +11,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+import javax.servlet.http.HttpSession;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -18,27 +20,45 @@ import java.util.List;
 public class UserController {
     @Autowired
     private UserService userService;
+    @Autowired
+    private AuthenticationService authenticationService;
+
+//    @GetMapping("")
+//    public String handleReq(HttpSession session) {
+//        if (!authenticationService.isLoggedIn(session)) {
+//            return "redirect:login";
+//        }
+//        return "redirect:/";
+//    }
 
     @GetMapping("/find")
-    public String findUserPageDisplay() {
-        return "user/search";
+    public String findUserPageDisplay(HttpSession session) {
+        if(authenticationService.isLoggedIn(session)){
+            return "user/search";
+        } else {
+            return "redirect:/";
+        }
     }
 
 
     @PostMapping("/find")
-    public String findUserByusername(@ModelAttribute("username") String username, Model model) {
-        User user = userService.getUser(username);
-        if (user != null) {
-            List<User> users = new ArrayList<>();
-            users.add(user);
-            model.addAttribute("users", users);
+    public String findUserByusername(@ModelAttribute("username") String username, Model model, HttpSession session) {
+        if(authenticationService.isLoggedIn(session)){
+            User user = userService.getUser(username);
+            if (user != null) {
+                List<User> users = new ArrayList<>();
+                users.add(user);
+                model.addAttribute("users", users);
+            }
+            return "user/all";
+        } else {
+            return "redirect:/";
         }
-        return "user/all";
     }
 
 
     @GetMapping("/register")
-    public String showRegistrationPage(Model model, final RedirectAttributes redirectAttributes) {
+    public String showRegistrationPage(Model model) {
         List<String> roles = new ArrayList<>();
         roles.add("USER");
         roles.add("ADMIN");
@@ -49,16 +69,14 @@ public class UserController {
 
 
     @PostMapping("/register")
-    public String addUser(@ModelAttribute("user") User user, @RequestParam("image") MultipartFile image, Model model, final RedirectAttributes redirectAttributes) {
+    public String addUser(@ModelAttribute("user") User user, @RequestParam("image") MultipartFile image, Model model,
+                          final RedirectAttributes redirectAttributes) {
         if (user.getRole() == null) {
             user.setRole("USER");
         }
         user = userService.addUser(user);
         if (user != null) {
             redirectAttributes.addFlashAttribute("success", "Please log in");
-            if(Helper.findLoggedInUserRole().equals("ADMIN")){
-                return "redirect:/user/list";
-            }
             return "redirect:/";
         } else {
             return "redirect:/user/register";
@@ -66,20 +84,40 @@ public class UserController {
     }
 
     @GetMapping("/list")
-    public String findAllUsers(Model model) {
-        List<User> list;
-        list = userService.getAllUsers();
-        model.addAttribute("users", list);
-        return "user/all";
+    public String findAllUsers(Model model, HttpSession session) {
+        if(authenticationService.isLoggedIn(session)){
+            List<User> list;
+            list = userService.getAllUsers();
+            model.addAttribute("users", list);
+            return "user/all";
+        } else {
+            return "redirect:/";
+        }
+
     }
 
     @GetMapping("/delete")
-    public String delete( @ModelAttribute("user") User user, final RedirectAttributes redirectAttributes) {
-        if (userService.delete(user.getId())) {
-            redirectAttributes.addFlashAttribute("success", "Successfully Deleted User..");
+    public String delete( @ModelAttribute("user") User user, final RedirectAttributes redirectAttributes, HttpSession session) {
+        if(authenticationService.isLoggedIn(session)){
+            if (userService.delete(user.getId())) {
+                redirectAttributes.addFlashAttribute("success", "Successfully Deleted User..");
+            } else {
+                redirectAttributes.addFlashAttribute("error", "Deletion Failed !");
+            }
+            return "redirect:list";
         } else {
-            redirectAttributes.addFlashAttribute("error", "Deletion Failed !");
+            return "redirect:/";
         }
-        return "redirect:list";
+
     }
+
+    @GetMapping("/about-us")
+    public String about(HttpSession session) {
+        if (authenticationService.isLoggedIn(session)){
+            return "about/about_us";
+        } else {
+            return "redirect:/";
+        }
+    }
+
 }
